+++ README.md
@@ -1,6 +1,6 @@
 # My To-do List App
 
-This is an application that helps users to organize and manage their tasks. Users can create a new task group and can also delete it if needed. Various tasks can be created within a particular task group. The title of the task, description, expected task completion date, and task priority is included during task creation. Each task can be edited or deleted if necessary. Double click on a completed task to highlight it. Tasks can also be sorted according to due-date and priority. For this project, JavaScript local storage is used to store the data on a user's computer. The implementation scripts are arranged in three different files consisting of similar operations and functions. This application is mobile responsive and can be viewed on various mobile devices. 
+This is an application that helps users to organize and manage their tasks. Users can create a new task group and can also delete it if needed. Various tasks can be created within a particular task group. This can be initiated by clicking on the rounded plus icon on the task display page to call up the required form. The title of the task, description, expected task completion date, and task priority must be included during task creation. Each task can be edited or deleted if necessary. Double click on a completed task to highlight it. Tasks can also be sorted according to due-date and priority. For this project, JavaScript local storage is used to store the data on a user's computer. The implementation scripts are arranged in two different files placed in the modules folder. They contain the application logic performing different operations and functions of that file. All the DOM manipulations are contained in the index.js file for clarity. This application is fully responsive and can be viewed on various mobile devices.
 
 ## Preview
